StructuredType move tests to *LatestIT
