+++ src/test/java/net/snowflake/client/jdbc/ResultSetIT.java
@@ -51,106 +51,6 @@ public class ResultSetIT extends ResultSet0IT {
     connection.close();
   }
 
-  public static class TestClass implements SQLData {
-
-    private String x;
-
-    @Override
-    public String getSQLTypeName() throws SQLException {
-      return null;
-    }
-
-    @Override
-    public void readSQL(SQLInput stream, String typeName) throws SQLException {
-      x = stream.readString();
-    }
-
-    @Override
-    public void writeSQL(SQLOutput stream) throws SQLException {}
-  }
-
-  @Test
-  public void testMapStructToObjectWithFactory() throws SQLException {
-    testMapJson(true);
-  }
-
-  @Test
-  public void testMapStructToObjectWithReflection() throws SQLException {
-    testMapJson(true);
-    testMapJson(false);
-    testMapArray(true);
-    testMapArray(false);
-    testReturnAsMap(true);
-    testReturnAsMap(false);
-  }
-
-  private void testMapJson(boolean registerFactory) throws SQLException {
-    if (registerFactory) {
-      SnowflakeObjectTypeFactories.register(TestClass.class, TestClass::new);
-    } else {
-      SnowflakeObjectTypeFactories.unregister(TestClass.class);
-    }
-    Connection connection = init();
-    Statement statement = connection.createStatement();
-    ResultSet resultSet = statement.executeQuery("select {'x':'a'}::OBJECT(x VARCHAR)");
-    resultSet.next();
-    TestClass object = resultSet.getObject(1, TestClass.class);
-    assertEquals("a", object.x);
-    statement.close();
-    connection.close();
-  }
-  private void testMapArray(boolean registerFactory) throws SQLException {
-    if (registerFactory) {
-      SnowflakeObjectTypeFactories.register(TestClass.class, TestClass::new);
-    } else {
-      SnowflakeObjectTypeFactories.unregister(TestClass.class);
-    }
-    Connection connection = init();
-    Statement statement = connection.createStatement();
-    ResultSet resultSet = statement.executeQuery("select [{'x':'aaa'},{'x': 'bbb'}]::ARRAY(OBJECT(x varchar))");
-    resultSet.next();
-    List<TestClass> objects = resultSet.unwrap(SnowflakeBaseResultSet.class).getList(1,  TestClass.class);
-    assertEquals(objects.get(0).x, "aaa");
-    assertEquals(objects.get(1).x, "bbb");
-    statement.close();
-    connection.close();
-  }
-
-  private void testReturnAsMap(boolean registerFactory) throws SQLException {
-    if (registerFactory) {
-      SnowflakeObjectTypeFactories.register(TestClass.class, TestClass::new);
-    } else {
-      SnowflakeObjectTypeFactories.unregister(TestClass.class);
-    }
-    Connection connection = init();
-    Statement statement = connection.createStatement();
-    ResultSet resultSet = statement.executeQuery("select {'x':{'x':'one'},'y':{'x':'two'},'z':{'x':'three'}}::MAP(VARCHAR, OBJECT(x VARCHAR));");
-    resultSet.next();
-    Map<String, TestClass> map = resultSet.unwrap(SnowflakeBaseResultSet.class).getMap(1,  TestClass.class);
-    assertEquals(map.get("x").x, "one");
-    assertEquals(map.get("y").x, "two");
-    assertEquals(map.get("z").x, "three");
-    statement.close();
-    connection.close();
-  }
-
-  @Test
-  public void testMapStructsFromChunks() throws SQLException {
-
-    Connection connection = init();
-    Statement statement = connection.createStatement();
-    ResultSet resultSet =
-        statement.executeQuery(
-            "select {'x':'a'}::OBJECT(x VARCHAR) FROM TABLE(GENERATOR(ROWCOUNT=>30000))");
-    int i = 0;
-    while (resultSet.next()) {
-      TestClass object = resultSet.getObject(1, TestClass.class);
-      assertEquals("a", object.x);
-    }
-    statement.close();
-    connection.close();
-  }
-
   @Test
   public void testGetColumnClassNameForBinary() throws Throwable {
     Connection connection = init();
