+++ src/test/java/net/snowflake/client/jdbc/ResultSetLatestIT.java
@@ -21,6 +21,7 @@ import net.snowflake.client.RunningOnGithubAction;
 import net.snowflake.client.category.TestCategoryResultSet;
 import net.snowflake.client.core.SFBaseSession;
 import net.snowflake.client.core.SessionUtil;
+import net.snowflake.client.core.structs.SnowflakeObjectTypeFactories;
 import net.snowflake.client.jdbc.telemetry.*;
 import net.snowflake.common.core.SFBinary;
 import org.apache.arrow.vector.Float8Vector;
@@ -927,4 +928,110 @@ public class ResultSetLatestIT extends ResultSet0IT {
       connection.close();
     }
   }
+
+  public static class TestClass implements SQLData {
+
+    private String x;
+
+    @Override
+    public String getSQLTypeName() throws SQLException {
+      return null;
+    }
+
+    @Override
+    public void readSQL(SQLInput stream, String typeName) throws SQLException {
+      x = stream.readString();
+    }
+
+    @Override
+    public void writeSQL(SQLOutput stream) throws SQLException {}
+  }
+
+  @Test
+  public void testMapStructToObjectWithFactory() throws SQLException {
+    testMapJson(true);
+  }
+
+  @Test
+  public void testMapStructToObjectWithReflection() throws SQLException {
+    testMapJson(true);
+    testMapJson(false);
+  }
+  @Test
+  public void testMapArrayToListWithReflection() throws SQLException {
+    testMapArrayAsList(true);
+    testMapArrayAsList(false);
+  }
+
+  @Test
+  public void testMapArrayWithReflection() throws SQLException {
+    testMapArray(true);
+    testMapArray(false);
+  }
+
+  private void testMapJson(boolean registerFactory) throws SQLException {
+    if (registerFactory) {
+      SnowflakeObjectTypeFactories.register(TestClass.class, TestClass::new);
+    } else {
+      SnowflakeObjectTypeFactories.unregister(TestClass.class);
+    }
+    Connection connection = init();
+    Statement statement = connection.createStatement();
+    ResultSet resultSet = statement.executeQuery("select {'x':'a'}::OBJECT(x VARCHAR)");
+    resultSet.next();
+    TestClass object = resultSet.getObject(1, TestClass.class);
+    assertEquals("a", object.x);
+    statement.close();
+    connection.close();
+  }
+  private void testMapArrayAsList(boolean registerFactory) throws SQLException {
+    if (registerFactory) {
+      SnowflakeObjectTypeFactories.register(TestClass.class, TestClass::new);
+    } else {
+      SnowflakeObjectTypeFactories.unregister(TestClass.class);
+    }
+    Connection connection = init();
+    Statement statement = connection.createStatement();
+    ResultSet resultSet = statement.executeQuery("select [{'x':'aaa'},{'x': 'bbb'}]::ARRAY(OBJECT(x varchar))");
+    resultSet.next();
+    List<TestClass> objects = resultSet.unwrap(SnowflakeBaseResultSet.class).getList(1,  TestClass.class);
+    assertEquals(objects.get(0).x, "aaa");
+    assertEquals(objects.get(1).x, "bbb");
+    statement.close();
+    connection.close();
+  }
+  private void testMapArray(boolean registerFactory) throws SQLException {
+    if (registerFactory) {
+      SnowflakeObjectTypeFactories.register(TestClass.class, TestClass::new);
+    } else {
+      SnowflakeObjectTypeFactories.unregister(TestClass.class);
+    }
+    Connection connection = init();
+    Statement statement = connection.createStatement();
+    ResultSet resultSet = statement.executeQuery("select [{'x':'aaa'},{'x': 'bbb'}]::ARRAY(OBJECT(x varchar))");
+    resultSet.next();
+    TestClass[] objects = resultSet.unwrap(SnowflakeBaseResultSet.class).getArray(1,  TestClass.class);
+    assertEquals(objects[0].x, "aaa");
+    assertEquals(objects[1].x, "bbb");
+    statement.close();
+    connection.close();
+  }
+
+  @Test
+  public void testMapStructsFromChunks() throws SQLException {
+
+    Connection connection = init();
+    Statement statement = connection.createStatement();
+    ResultSet resultSet =
+            statement.executeQuery(
+                    "select {'x':'a'}::OBJECT(x VARCHAR) FROM TABLE(GENERATOR(ROWCOUNT=>30000))");
+    int i = 0;
+    while (resultSet.next()) {
+      TestClass object = resultSet.getObject(1, TestClass.class);
+      assertEquals("a", object.x);
+    }
+    statement.close();
+    connection.close();
+  }
+
 }
